model: Increase precision of lightmap extents calculations
This bug is a counterpart to this bug in quakespasm:
   https://sourceforge.net/p/quakespasm/patches/15/

If the precision of the calculation in WorldToTexCoord done by ‘light’
is significantly different than the precision of the same calculation
in the engine, you can get corrupted lightmaps. The problem affects
light when it’s built as something other than vanilla 32-bit
x86. (e.g. x86_64 on a mac).

+++ common/model.c
@@ -922,10 +922,27 @@ CalcSurfaceExtents(const brushmodel_t *brushmodel, msurface_t *surf)
 	    const medge_t *const edge = &brushmodel->edges[-edgenum];
 	    vertex = &brushmodel->vertexes[edge->v[1]];
 	}
+
+	/*
+	 * The (long double) casts below are important: The original code was
+	 * written for x87 floating-point which uses 80-bit floats for
+	 * intermediate calculations. But if you compile it without the casts
+	 * for modern x86_64, the compiler will round each intermediate result
+	 * to a 32-bit float, which introduces extra rounding error.
+	 *
+	 * This becomes a problem if the rounding error causes the light
+	 * utilities and the engine to disagree about the lightmap size for
+	 * some surfaces.
+	 *
+	 * Casting to (long double) keeps the intermediate values at at least
+	 * 64 bits of precision, probably 128.
+	 */
 	for (j = 0; j < 2; j++) {
-	    val = vertex->position[0] * texinfo->vecs[j][0] +
-		vertex->position[1] * texinfo->vecs[j][1] +
-		vertex->position[2] * texinfo->vecs[j][2] + texinfo->vecs[j][3];
+	    val =
+		(long double)vertex->position[0] * texinfo->vecs[j][0] +
+		(long double)vertex->position[1] * texinfo->vecs[j][1] +
+		(long double)vertex->position[2] * texinfo->vecs[j][2] +
+		                                   texinfo->vecs[j][3];
 	    if (val < mins[j])
 		mins[j] = val;
 	    if (val > maxs[j])
